Fix noise scaling for pre-normalized gradients
noise_std was σ×C_internal (2048x too large) because gradient was
pre-divided by num_tokens×num_microbatches but noise wasn't.
Fix: noise_std = σ × C_per_token / num_microbatches when
--dp-clipping-norm-per-token is set.

Also: fix transformers version check in loader_llama_mistral.py,
add TE native norm classes to ghost_clipping.py _NORM_CLASSES,
fix _ReplayableIterator rewind for TP>1.

+++ megatron/core/pipeline_parallel/schedules.py
@@ -632,7 +632,7 @@ def _dp_sgd_ghost_forward_backward(
                 continue
             if p.grad is not None and hasattr(p, 'main_grad') and p.main_grad is not None:
                 if not getattr(p, 'grad_added_to_main_grad', False):
-                    p.main_grad.add_(p.grad.data.float())
+                    p.main_grad.add_(p.grad.data)  # FP16→FP32 auto-promoted by PyTorch
                 p.grad = None
 
     # ===== AFTER ALL MICROBATCHES: finalize =====
